% Auto-generated: do not edit by hand
\name{dashClustergrammer}

\alias{dashClustergrammer}

\title{DashClustergrammer component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
dashClustergrammer(id=NULL, label=NULL, value=NULL, network_data=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{label}{Character. A label that will be printed when this component is rendered.}

\item{value}{Character. The value displayed in the input.}

\item{network_data}{Named list. The JSON data that Clustergrammer takes in as input}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

